Generate list of disks based on Model name
This is necessary for now as the SCSI IDs keep changing on me on this
system. I'll likely change this to search by something more reliable
than the device model, but this will get me started and running for now,
it is at least reletively reliable and consistent (and stable).